Cache computation of initial resource meta in plugin
