+++ packages/resourceful-redux/src/request-statuses-plugin/create.js
@@ -1,24 +1,14 @@
 import actionTypes from '../action-types';
 import reducerGenerator from '../utils/reducer-generator';
 import cruReducerHelper from '../utils/cru-reducer-helper';
-import initialResourceMetaState from '../utils/initial-resource-meta-state';
 import requestStatuses from '../utils/request-statuses';
 
 const create = reducerGenerator('create', requestStatuses.PENDING);
 const createFail = reducerGenerator('create', requestStatuses.FAILED);
 const createNull = reducerGenerator('create', requestStatuses.NULL)
 
-function createSucceed(state, action, options = {}) {
-  // Could be cached earlier on!
-  const customInitialMeta = options.initialResourceMeta || {};
-  const optionsToSend = {
-    initialResourceMeta: {
-      ...initialResourceMetaState,
-      ...customInitialMeta
-    }
-  };
-
-  return cruReducerHelper(state, action, optionsToSend, {
+function createSucceed(state, action, options) {
+  return cruReducerHelper(state, action, options, {
     readStatus: requestStatuses.SUCCEEDED,
     createStatus: requestStatuses.SUCCEEDED,
   });
