+++ packages/resourceful-redux/src/request-statuses-plugin/read.js
@@ -1,22 +1,13 @@
 import reducerGenerator from '../utils/reducer-generator';
 import cruReducerHelper from '../utils/cru-reducer-helper';
-import initialResourceMetaState from '../utils/initial-resource-meta-state';
 import requestStatuses from '../utils/request-statuses';
 
 const read = reducerGenerator('read', requestStatuses.PENDING);
 const readFail = reducerGenerator('read', requestStatuses.FAILED);
 const readNull = reducerGenerator('read', requestStatuses.NULL);
 
-function readSucceed(state, action, options = {}) {
-  const customInitialMeta = options.initialResourceMeta || {};
-  const optionsToSend = {
-    initialResourceMeta: {
-      ...initialResourceMetaState,
-      ...customInitialMeta
-    }
-  };
-
-  return cruReducerHelper(state, action, optionsToSend, {
+function readSucceed(state, action, options) {
+  return cruReducerHelper(state, action, options, {
     readStatus: requestStatuses.SUCCEEDED
   });
 }
